Chips are moveable but not keeping their new positions
