--- client/components/sidePanel/categoryFilter/Categories.js
@@ -1,13 +1,0 @@
-import React from 'react'
-import {SelectedCategories, CategoryFilter} from '../..'
-
-const Categories = props => {
-  return (
-    <div className="categories-display side-panel-body">
-      <CategoryFilter />
-      <SelectedCategories />
-    </div>
-  )
-}
-
-export default Categories
