--- client/components/sidePanel/categoryFilter/compareButton.js
@@ -1,30 +1,0 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import {withStyles} from '@material-ui/core/styles'
-import Button from '@material-ui/core/Button'
-
-const styles = theme => ({
-  button: {
-    margin: theme.spacing.unit
-  },
-  input: {
-    display: 'none'
-  }
-})
-
-function CompareButton(props) {
-  const {classes} = props
-  return (
-    <div>
-      <Button variant="outlined" color="primary" className={classes.button}>
-        Compare My Homes
-      </Button>
-    </div>
-  )
-}
-
-CompareButton.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(CompareButton)
