+++ client/index.js
@@ -8,7 +8,7 @@ import {store, persistor} from './store'
 import App from './app'
 
 // establishes socket connection
-import './socket'
+// import './socket'
 
 ReactDOM.render(
   <Provider store={store}>
