+++ server/socket/index.js
@@ -1,9 +1,9 @@
-module.exports = io => {
-  io.on('connection', socket => {
-    console.log(`A socket connection to the server has been made: ${socket.id}`)
+// module.exports = io => {
+//   io.on('connection', socket => {
+//     console.log(`A socket connection to the server has been made: ${socket.id}`)
 
-    socket.on('disconnect', () => {
-      console.log(`Connection ${socket.id} has left the building`)
-    })
-  })
-}
+//     socket.on('disconnect', () => {
+//       console.log(`Connection ${socket.id} has left the building`)
+//     })
+//   })
+// }
